Improve exception handling during database connect
If the settings file cannot be read (or some other error occurs) during
the sqlite connect, this resulted in an exception which was caught.
The finally block then attempted to close the connection which may not
have been created which triggered a second exception. This could mask
the real error.

- Catch and log failure to connect to the database
- Catch and log of closing the connection fails.